refactor: SlippWebApplicationInitializer에서 ApplicationContext에 bean들을 등록하도록 수정
