+++ slipp/src/main/java/slipp/NotFoundBeanException.java
@@ -1,4 +1,4 @@
-package slipp.support.context;
+package slipp;
 
 public class NotFoundBeanException extends RuntimeException {
     private static final String MESSAGE = "해당 Bean이 존재하지 않습니다. : ";
