+++ slipp/src/main/java/slipp/dao/UserDao.java
@@ -2,8 +2,8 @@ package slipp.dao;
 
 import nextstep.jdbc.JdbcTemplate;
 import nextstep.jdbc.RowMapper;
+import slipp.ApplicationContext;
 import slipp.domain.User;
-import slipp.support.context.ApplicationContext;
 
 import java.util.List;
