+++ slipp/src/main/java/slipp/domain/User.java
@@ -8,7 +8,8 @@ public class User {
     private String name;
     private String email;
 
-    public User() {}
+    public User() {
+    }
 
     public User(String userId, String password, String name, String email) {
         this.userId = userId;
