+++ slipp/src/test/java/support/test/TestInitializer.java
@@ -1,6 +1,7 @@
 package support.test;
 
 import org.springframework.mock.web.MockServletContext;
+import slipp.ApplicationContextInitializer;
 import slipp.support.context.ContextLoaderListener;
 
 import javax.servlet.ServletContextEvent;
@@ -8,6 +9,8 @@ import javax.servlet.ServletContextListener;
 
 public class TestInitializer {
     public static void initialize() {
+        ApplicationContextInitializer.registerBeans();
+
         MockServletContext servletContext = new MockServletContext();
         ServletContextListener listener = new ContextLoaderListener();
         ServletContextEvent event = new ServletContextEvent(servletContext);
